Updated Telize to it's new paid API
